[REF] 패키지 분리 완료 및 onClick 내부 Intent flags 설정 미세 수정
