+++ app/src/main/java/com/example/letssopt/ui/signup/SignUpActivity.kt
@@ -1,21 +1,21 @@
-package com.example.letssopt
+package com.example.letssopt.ui.signup
 
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
+import android.util.Patterns
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardActions
@@ -23,11 +23,15 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -41,28 +45,44 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.letssopt.ui.components.DefaultButton
 import com.example.letssopt.ui.components.DefaultTextField
+import com.example.letssopt.ui.login.LoginActivity
 import com.example.letssopt.ui.theme.*
 import com.example.letssopt.ui.theme.LETSSOPTTheme
+import kotlinx.coroutines.launch
 
-class LoginActivity : ComponentActivity() {
+class SignUpActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        Log.d("LoginActivity", "onCreate: ${intent.getStringExtra("id")} and ${intent.getStringExtra("pw")}")
-        var tempId = intent.getStringExtra("id")
-        var tempPw = intent.getStringExtra("pw")
-
         setContent {
             LETSSOPTTheme {
-                Scaffold(modifier = Modifier
-                    .fillMaxSize()) { innerPadding ->
-                    LoginContent(
-                        modifier = Modifier
-                            .padding(innerPadding)
-                            .consumeWindowInsets(innerPadding),
-                        saveId = tempId,
-                        savePw = tempPw
-                    )
+                val scope = rememberCoroutineScope()
+                val snackbarHostState = remember { SnackbarHostState() }
+
+                Scaffold(
+                    snackbarHost = {
+                        SnackbarHost(
+                            modifier = Modifier
+                                .navigationBarsPadding()
+                                .imePadding(),
+                            hostState = snackbarHostState
+                    ) {
+                        Snackbar(
+                            snackbarData = it,
+                            containerColor = AsSurface,
+                            contentColor = AsWhite
+                        )
+                    } },
+                    modifier = Modifier.fillMaxSize()) { innerPadding ->
+                        SignupContent(
+                            modifier = Modifier.padding(innerPadding)
+                                .consumeWindowInsets(innerPadding),
+                            onShowSnack = { message ->
+                                scope.launch {
+                                    snackbarHostState.showSnackbar(message)
+                                }
+                            }
+                        )
                 }
             }
         }
@@ -70,14 +90,18 @@ class LoginActivity : ComponentActivity() {
 }
 
 @Composable
-fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?) {
+fun SignupContent(
+    modifier: Modifier = Modifier,
+    onShowSnack: (String) -> Unit
+) {
     val context = LocalContext.current
+
     val focusManager = LocalFocusManager.current
     val scrollState = rememberScrollState()
-    val interactionSource = remember { MutableInteractionSource() }
 
     var textId by remember { mutableStateOf("") }
     var textPw by remember { mutableStateOf("") }
+    var textPwCheck by remember { mutableStateOf("") }
 
     //전반적인 레이아웃
     Column(modifier = modifier
@@ -90,8 +114,8 @@ fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?
         //위쪽 컴포넌트 정렬을 위한 2차 레이아웃 wrapper
         Column(modifier = Modifier
             .fillMaxWidth()
-            .weight(1f)
-            .verticalScroll(scrollState),
+            .verticalScroll(scrollState)
+            .weight(1f),
             horizontalAlignment = Alignment.Start,
             verticalArrangement = Arrangement.Top
             ) {
@@ -100,18 +124,17 @@ fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?
                 modifier = Modifier
                     .padding(top = 40.dp)
                     .align(Alignment.CenterHorizontally),
-                style = MaterialTheme.typography.titleLarge.copy(fontSize = 36.sp, color = AsPrimary),
-                text = "watcha",
-                color = AsPrimary
-                )
+                style = MaterialTheme.typography.titleLarge.copy(fontSize = 36.sp ,color = AsPrimary),
+                text = "watcha"
+            )
 
             Text(
                 modifier = Modifier
                     .padding(top = 20.dp),
                 style = MaterialTheme.typography.titleLarge.copy(fontSize = 20.sp),
-                text = "이메일로 로그인",
+                text = "회원가입",
                 color = AsWhite
-                )
+            )
             //텍스트필드 섹션
             Text(
                 modifier = Modifier
@@ -139,13 +162,10 @@ fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?
                     onNext = { focusManager.moveFocus(FocusDirection.Down) }
                 )
             )
-
-            Text(
-                modifier = Modifier
-                    .padding(top = 12.dp),
-                style = MaterialTheme.typography.labelSmall,
+            Text(style = MaterialTheme.typography.labelSmall,
                 text = "비밀번호",
-                color = AsSecondaryText
+                color = AsSecondaryText,
+                modifier = Modifier.padding(top = 12.dp)
             )
 
             DefaultTextField(
@@ -160,6 +180,33 @@ fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?
                 tfVisible = false,
                 keyboardOptions = KeyboardOptions(
                     keyboardType = KeyboardType.Password,
+                    imeAction = ImeAction.Next
+                ),
+                keyboardActions = KeyboardActions(
+                    onNext = { focusManager.moveFocus(FocusDirection.Down) }
+                )
+            )
+
+            Text(
+                modifier = Modifier
+                    .padding(top = 12.dp),
+                style = MaterialTheme.typography.labelSmall,
+                text = "비밀번호 확인",
+                color = AsSecondaryText
+            )
+
+            DefaultTextField(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 4.dp),
+                text = textPwCheck,
+                onValueChange = { newText ->
+                    textPwCheck = newText
+                },
+                hint = "비밀번호를 다시 입력하세요",
+                tfVisible = false,
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Password,
                     imeAction = ImeAction.Done
                 ),
                 keyboardActions = KeyboardActions(
@@ -167,79 +214,64 @@ fun LoginContent(modifier: Modifier = Modifier, saveId: String?, savePw: String?
                 )
             )
         }
-
-        //여기 두번쨰 컬럼에서 지정하지도 않은 padding이 양쪽 너비에 들어가서 로그인 버튼이 이상한 패딩이 적용됨
-        Column(modifier = Modifier
-            .fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Bottom
-        ) {
-
-            Text(modifier = Modifier
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = {
-                        val intent = Intent(context, SignUpActivity::class.java)
-                        context.startActivity(intent)
-                    }
-                )
-                .padding(bottom = 10.dp, top = 10.dp),
-                style = MaterialTheme.typography.labelSmall,
-                color = AsSecondaryText,
-                text = "아직 계정이 없으신가요? 회원가입"
-
-            )
-
-            DefaultButton(modifier = Modifier
+        // 이쪽 부분에선 래핑해줬던 레이아웃 제거
+        DefaultButton(
+            modifier = Modifier
                 .fillMaxWidth()
-                .padding(bottom = 10.dp),
-                text = "로그인",
-                onClick = {
-                    val intent2 = Intent(context, MainActivity::class.java).apply {
-                        flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
-                    }
-
-                    validateLogin(
-                        context = context,
-                        saveId = saveId,
-                        savePw = savePw,
-                        textId = textId,
-                        textPw = textPw,
-                        successIntent = intent2
-                    )
+                .padding(bottom = 10.dp, top = 20.dp),
+            "회원가입",
+            onClick = {
+                val intent = Intent(context, LoginActivity::class.java).apply {
+                    flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
                 }
-            )
-        }
+
+                validateSignUp(
+                    context = context,
+                    textId = textId,
+                    textPw = textPw,
+                    textPwCheck = textPwCheck,
+                    successIntent = intent,
+                    onShowSnack = onShowSnack
+                )
+
+        },
+            btEnabled = textId.isNotEmpty() &&
+                    textPw.isNotEmpty() &&
+                    textPwCheck.isNotEmpty()
+        )
     }
 }
 
-private fun validateLogin(
-    context: android.content.Context,
-    saveId: String?,
-    savePw: String?,
+private fun validateSignUp(
+    context: Context,
     textId: String,
     textPw: String,
-    successIntent: android.content.Intent
+    textPwCheck: String,
+    successIntent: Intent,
+    onShowSnack: (String) -> Unit
 ) {
-    if (saveId == null || savePw == null) {
-        Toast.makeText(context, "회원가입이 필요합니다.", Toast.LENGTH_LONG).show()
+    if (Patterns.EMAIL_ADDRESS.matcher(textId).matches() &&
+        textPw.length in 8..12 && // 💡 코틀린 스타일로 더 깔끔하게 다듬었습니다.
+        textPw == textPwCheck
+    ) {
+        Toast.makeText(context, "회원가입 성공!", Toast.LENGTH_SHORT).show()
+
+        successIntent.putExtra("id", textId)
+        successIntent.putExtra("pw", textPw)
+
+        context.startActivity(successIntent)
     } else {
-        if (saveId == textId && savePw != textPw) {
-            Toast.makeText(context, "비밀번호가 틀렸습니다. 다시 확인해주세요.", Toast.LENGTH_SHORT).show()
-        } else if (saveId == textId && savePw == textPw) {
-            context.startActivity(successIntent)
-            Toast.makeText(context, "로그인 성공. 환영합니다. $textId 님.", Toast.LENGTH_SHORT).show()
-        } else {
-            Toast.makeText(context, "로그인 실패. 다시 시도해주세요", Toast.LENGTH_SHORT).show()
-        }
+        onShowSnack("회원가입에 실패했습니다. 올바른 정보를 입력해주세요.")
     }
 }
+
 
 @Preview(showBackground = true)
 @Composable
-private fun LoginContentPreview() {
+private fun SignupContentPreview() {
     LETSSOPTTheme {
-        LoginContent(saveId = "id", savePw = "pw")
+        SignupContent {
+
+        }
     }
 }
